% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insp.plot.R
\name{ask.reveal.data}
\alias{ask.reveal.data}
\title{Asks whether actual datum plot sould be shown.}
\usage{
ask.reveal.data(real.data, ask = TRUE)
}
\arguments{
\item{real.data}{plot position with real data}

\item{ask}{Whether ask before showing real data plot}
}
\description{
This function is used internally.
}
