% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insp.plot.R
\name{extract.data}
\alias{extract.data}
\title{Extract model data}
\usage{
extract.data(model)
}
\arguments{
\item{model}{A fitted model}
}
\value{
Model data.
}
\description{
Extract model data from `lm()`, `glm()`, `lmer()`,
`lme()` and `gls()` models.
}
