% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insp.plot.R
\name{extract.resid}
\alias{extract.resid}
\title{Extract model residuals}
\usage{
extract.resid(model, bin.size = 10)
}
\arguments{
\item{model}{A fitted model}

\item{bin.size}{For glm binomial family only. Binned residuals.}
}
\value{
Model specific residuals.
}
\description{
Extract model specific results from `lm()`, `glm()`, `lmer()`,
`lme()` and `gls()` models.
}
