% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insp.plot.R
\name{plots.acf.r}
\alias{plots.acf.r}
\title{Plot residuals auto-correlation estimates
from `insp.sim()` data.}
\usage{
plots.acf.r(models, ask = TRUE)
}
\arguments{
\item{models}{list with models from `insp.sim()` function.}

\item{ask}{whether ask before showing real data plot.}
}
\description{
This function depends on the order of the data.
}
