% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insp.plot.R
\name{plots.cook.r}
\alias{plots.cook.r}
\title{Plot residuals cook's distances from `insp.sim()` data.}
\usage{
plots.cook.r(models, ask = TRUE)
}
\arguments{
\item{models}{list with models from `insp.sim()` function.}

\item{ask}{whether ask before showing real data plot.}
}
\description{
Plot residuals cook's distances from `insp.sim()` data.
}
