% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insp.plot.R
\name{plots.hist.r}
\alias{plots.hist.r}
\title{Plot residuals histograms from `insp.sim()` data.}
\usage{
plots.hist.r(resid, ask = TRUE)
}
\arguments{
\item{resid}{list with models residuals}

\item{ask}{whether ask before showing real data plot}
}
\description{
Plot residuals histograms from `insp.sim()` data.
}
