% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insp.plot.R
\name{plots.insp.sim}
\alias{plots.insp.sim}
\title{Plot simulated data}
\usage{
plots.insp.sim(insp.sim, resid, fitted, data, pred, ask = TRUE)
}
\arguments{
\item{insp.sim}{A fitted model simulation list form `insp.sim`}

\item{resid}{A list of model simulation residuals}

\item{fitted}{A list of model simulation fitted data}

\item{data}{A list of model simulation data}

\item{pred}{vector with colnames of predictor variables}

\item{ask}{Whether ask before showing real data plot}
}
\description{
Plot a series of validation plots based on results from
`insp.sim()` function.
}
\examples{
n<-100
pred1<-rnorm(n=n, mean=0, sd=1)
pred2<-rnorm(n=n, mean=1, sd=2)
pred3<-gl(n=5,k=n/5)
ERROR<-rnorm(n=n, mean=0, sd=2)
a<-1
b<-2
c<-rep(runif(n=length(levels(pred3)),min = -2, max = 2),
       each=n/length(levels(pred3)))
Y <- a*pred1 + b*pred2 + c + ERROR
lm<-lm(Y ~ pred1 + pred2 + pred3)
sim<-insp.sim(model=lm)
plots.insp.sim(insp.sim=sim,
              resid=lapply(sim,extract.resid),
              fitted=lapply(sim,fitted),
              data=lapply(sim,extract.data),
              pred = c("pred1","pred2","pred3"),
              ask = FALSE)
}
