% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insp.plot.R
\name{plots.qqnorm.random.r}
\alias{plots.qqnorm.random.r}
\title{Plot random residuals from `insp.sim()` mixed-model data.}
\usage{
plots.qqnorm.random.r(resid, ask = TRUE)
}
\arguments{
\item{resid}{list with model residuals}

\item{ask}{whether ask before showing real data plot}
}
\description{
Plot random residuals from `insp.sim()` mixed-model data.
}
