% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insp.plot.R
\name{plots.resid.bin.r}
\alias{plots.resid.bin.r}
\title{Plot residuals from `insp.sim()` binomial data.}
\usage{
plots.resid.bin.r(resid, ask = TRUE)
}
\arguments{
\item{resid}{list with models residuals}

\item{ask}{whether ask before showing real data plot}
}
\description{
Plot residuals from `insp.sim()` binomial data.
}
