% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insp.plot.R
\name{plots.resid.x.pred.r}
\alias{plots.resid.x.pred.r}
\title{Plot residuals x all covariates from `insp.sim()` data.}
\usage{
plots.resid.x.pred.r(resid, data, pred, ask = TRUE)
}
\arguments{
\item{resid}{list with model residuals}

\item{data}{list with model data}

\item{pred}{vector with colnames of predictor variables}

\item{ask}{whether ask before showing real data plot}
}
\description{
Plot residuals x all covariates from `insp.sim()` data.
}
