% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insp.plot.R
\name{random.plots}
\alias{random.plots}
\title{Generates a plot random order sequence.}
\usage{
random.plots(n = 20, datum = 20)
}
\arguments{
\item{n}{number of plots}

\item{datum}{plot number with real data}
}
\description{
This function is used internally.
}
